requirements.txt
Flask==3.0.0
google-genai>=0.3.0
python-dotenv==1.0.0
Werkzeug==3.0.1
supabase==2.27.2

